+++ src/main/java/be15fintomatokatchupbe/relation/service/PipeUserService.java
@@ -4,6 +4,7 @@ import be15fintomatokatchupbe.campaign.command.domain.aggregate.entity.Pipeline;
 import be15fintomatokatchupbe.relation.domain.PipelineUser;
 import be15fintomatokatchupbe.relation.repository.PipeUserRepository;
 import be15fintomatokatchupbe.user.command.application.service.UserCommendService;
+import be15fintomatokatchupbe.user.command.application.support.UserHelperService;
 import be15fintomatokatchupbe.user.command.domain.aggregate.User;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,11 +14,11 @@ import java.util.List;
 @Service
 @AllArgsConstructor
 public class PipeUserService {
-    private final UserCommendService userCommendService;
+    private final UserHelperService userHelperService;
     private final PipeUserRepository pipeUserRepository;
 
     public void saveUserList(List<Long> userList, Pipeline pipeline){
-        List<User> foundUserList = userCommendService.findValidUserList(userList);
+        List<User> foundUserList = userHelperService.findValidUserList(userList);
 
         List<PipelineUser> pipelineUserList = foundUserList.stream().map(user ->
             PipelineUser.builder()
