+++ src/main/java/be15fintomatokatchupbe/user/command/application/service/UserCommendService.java
@@ -62,18 +62,4 @@ public class UserCommendService {
         user.setPassword(passwordEncoder.encode(request.getNewPassword()));
         userRepository.save(user);
     }
-
-
-    /* 참조용 */
-    
-    // 삭제되지 않은 유저 찾기
-    public User findValidUser(Long id){
-        return userRepository.findByUserIdAndIsDeleted(id, StatusType.N)
-                .orElseThrow(() -> new BusinessException(UserErrorCode.USER_NOT_FOUND));
-    }
-
-    // 유저 목록으로 찾기
-    public List<User> findValidUserList(List<Long> userId){
-        return userRepository.findByUserIdInAndIsDeleted(userId, StatusType.N);
-    }
 }
